Add each source file individually
Previously we used a glob to search for all our c files en-mass. This
has a few downsides when compared to populating a source list manually.
It limits our ability to split up the program into internal libraries
because we do not have fine-grained control of what the exact source
files are. In the same way, if we want to include whole external
dependencies using git submodule, we need to be able to exclude certain
c files (e.g. tests which break the build). Additionally, globbing for
files requires you re-run cmake whenever new files are added (via git or
otherwise). If they are listed individually, cmake will detect a change
in CMakeLists.txt and only rerun when that has changed. See the
commentary at
<https://cmake.org/cmake/help/latest/command/aux_source_directory.html>
for a more thorough explanation.

With regard to the individual CMakeLists.txt files added in this commit,
all paths need to be absolute (hence the gratuitous use of
CMAKE_CURRENT_SOURCE_DIR), but this may change in the future, as the
build is broken up in a more logical fasion.