Split build into several smaller static libraries
Storing all sources in a SOURCES variable works, but makes modifying the
build system relatively brittle. We cannot (for example) relocate the
declaration of the main executable from the top-level CMakeLists.txt
easily. It is also difficult to set up optional use of installed
libraries included in the external dir. Additionally, it allows us to
get rid of some uses of CMAKE_CURRENT_SOURCE_DIR (but not all).

Libraries have been created corresponding to the directories in
src/[Client|Shared]/. Currently, src/Shared/External is not organized in
this manner, but will be looked at in a future update. One upside of
this organization is that inter-directory dependencies are now obvious
(as they result in linking errors if left unsatisfied). The current
dependency DAG is

RankCheck
	system
		graphics
		gui3
			utils
	config
	content