Switched to using submodule or system version of RapidJSON
RapidJSON is a package in some linux distributions, so we should use the
system version when it is installed and the version from git otherwise.
To do this, we create an INTERFACE target which sets the appropriate
include directories when it is included. Add it as a target wherever the
header files are needed (RankCheck, system, and config at the moment)