Switch to using submodule/system PugiXML
This is similar to previous versions, except we now need to build an
external project if pugixml isn't found on the system. FindPugi defines
a pugixml target, which makes our life a bit easier(-ish). We still need
to define it globally, so we need to do a bit of an ugly hack on cmake
versions prior to 3.11 (aka almost all of them as it was released 3 days
ago). Unfortunately we can't undefine our overridden function, as it
would cause an infinite loop (see
https://cmake.org/pipermail/cmake/2011-March/043320.html).