Store player database in cache directory
Instead of trying to store the player database in the same directory as
the executable (which may not work, e.g. when installed on linux or
windows), place it in the user's cache directory. On Linux the file will
usually be `~/.cache/rankcheck/players.db`, and on Windows it will usually
be `%APPDATA%\rankcheck\players.db`. Fixes #4.